+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - DefaultTree: Time tool opens again by changing WMS-Time layer
 - Issue #821: Sorting the table in featureLister tool now works again
 - Issue #845: The buffer analysis tool now shows results correctly again
+- Fixed a bug that caused some layers in the topic tree to disappear from the typ "default"
 
 ---